[fix: login]

+++ src/main/java/com/swd391/bachhoasi/repository/AdminRepository.java
@@ -4,8 +4,10 @@ import com.swd391.bachhoasi.model.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, BigDecimal> {
 
     Admin findByUsername(String username);
 }
